Reset data-person-index when editing a HH member
This commit fixes a bug I encounter when trying to edit the Head Of
Household. Basically, when editing the HOH the `data-person-index`
attribute was set to 2 and did not reset when filling the screener on
edit.

This led to not being able to continue to the recap page after step 10
with the error:
```
Type Error e is undefined
```

The undefined error was trigger when iterating over the `this._people`
array. Since the data-person-index was set to 2, the `this._people`
array at index 1 was set to undefined.
```
[person, undefined, person]
```

The fix is to reset the attribute data-person-index on edit so that
`this._people` array was set at the correct index.

This were a fun couple of days trying to fix this one! :cheers: