Reset HOH to element at index 0 in people array
Currently when you fill the screener and set yourself as not the Head Of
Hosehold and then edit yourself to be HOH the screener recap will
this play 2 Head Of Households.

There can only be one HOH. This commit will fix this bug by reseting the
member in the `this._people` to be HOH at index 0 when editing yourself
to be HOH. This is done by iterating over the `this._people`,
removing the element from the array that is set to be HOH and is not
elemet at index 0.

When you remove other elements from the array at this point it doesn't
matter because as you edit the screener application the new members will
be added to the array. Thus, displaying the correct HOH and correct
household members.

Resolves: AC-1201 https://hhsconnect.atlassian.net/browse/AC-1201