Head Of Household Bug Fixes (#148)
* Reset HOH to element at index 0 in people array

Currently when you fill the screener and set yourself as not the Head Of
Hosehold and then edit yourself to be HOH the screener recap will
this play 2 Head Of Households.

There can only be one HOH. This commit will fix this bug by reseting the
member in the `this._people` to be HOH at index 0 when editing yourself
to be HOH. This is done by iterating over the `this._people`,
removing the element from the array that is set to be HOH and is not
elemet at index 0.

When you remove other elements from the array at this point it doesn't
matter because as you edit the screener application the new members will
be added to the array. Thus, displaying the correct HOH and correct
household members.

* Set personIndex to 0 when editing HOH

When writting to the attribute `data-person-index` after editing the HOH
the attribute is not reset. The data `personIndex` has to be reset in
order to start at index 0 when adding new household members.