Decode URL to make unicode characters valid UTF-8 (#150)
It seems somewhere along the line the slug use in the id for the toggle
component got encoded. The urlencoded %e2%80%8b string represents
U+E2808B which is a zero-width-space. This is the reason why the toggle
function wouldn't work in the locations page.

The solution is to use PHP method urldecode to decode the slug. Thus,
converting the unicode into valid UTF-8. This was a sneaky bug.

** related
https://stackoverflow.com/questions/60157884/twig-how-to-interpolate-within-json?noredirect=1#comment106402222_60157884

https://www.php.net/manual/en/function.urldecode.php