Add Alert field to locations,Programs and Pages
This commit makes it possible to display an Alert in a single page,
location and program. It also makes it possible to display an alert
along a sitewide alert without conflicting. Meaning that an alert on a
page, program or location and be different than the sitewide alert, thus
displaying both at the same time.