Reduce latency to requests by using DNS prefetch
consider pairing dns-prefetch with the preconnect hint. While
dns-prefetch only performs a DNS lookup, preconnect establishes a
connection to a server. This process includes DNS resolution, as well as
establishing the TCP connection, and performing the TLS handshake—if a
site is served over HTTPS. Combining the two provides an opportunity to
further reduce the perceived latency of cross-origin requests. You can
safely use them together

Note: If a page needs to make connections to many third-party domains,
preconnecting them all is counterproductive. The preconnect hint is best
used for only the most critical connections.
For the others, just use <link rel="dns-prefetch">
to save time on the first step — the DNS lookup.